+++ docs/snippets/flipxy.ts
@@ -1,4 +1,4 @@
-import { Toodle } from "@bloop.gg/toodle";
+import { Toodle } from "@bloopjs/toodle";
 
 const canvas = document.querySelector("canvas")!;
 const toodle = await Toodle.attach(canvas, { filter: "nearest" });