+++ examples/15-text-layer.ts
@@ -1,4 +1,4 @@
-import { Toodle } from "@bloop.gg/toodle";
+import { Toodle } from "@bloopjs/toodle";
 import { createCanvas } from "./util";
 
 const canvas = createCanvas(window.innerWidth, window.innerHeight);